Count CalWORKs excess vehicle equity toward the resource limit
Per CDSS ACLs 25-37 and 26-38 (MPP 42-215.4), vehicle value above the
CalWORKs vehicle limit counts toward the maximum resource limit rather
than disqualifying the family outright. Replace the hard eligibility
cliff (ca_tanf_vehicle_value_eligible) with a countable excess variable
(ca_tanf_countable_vehicle_value) that flows into ca_tanf_resources.

Closes #9032
