+++ changelog.d/ca-calworks-vehicle-excess-equity.changed.md
@@ -1,0 +1,1 @@
+Count CalWORKs vehicle value above the vehicle limit toward the resource limit instead of denying eligibility outright, per CDSS ACLs 25-37 and 26-38.
