+++ policyengine_us/parameters/gov/states/ca/cdss/tanf/cash/resources/sources.yaml
@@ -7,7 +7,10 @@ metadata:
   reference:
     - title: Los Angeles Department of Public Social Services E-Policy | Property Definitions
       href: https://epolicy.dpss.lacounty.gov/epolicy/epolicy/server/general/projects_responsive/ePolicyMaster/index.htm?&area=general&type=responsivehelp&ctxid=&project=ePolicyMaster#t=mergedProjects%2FCalWORKs%2FCalWORKs%2F42-200_Property%2F42-200_Property.htm%23Definitionsbc-4&rhtocid=_3_1_2_0_3
+    - title: California Department of Social Services, All County Letter No. 26-38, Changes to Vehicle Value Limits
+      href: https://www.cdss.ca.gov/Portals/9/Additional-Resources/Letters-and-Notices/ACLs/2026/26-38.pdf#page=2
 values:
   2022-07-01:
     - assessed_property_value # real_property
     - spm_unit_cash_assets # personal_property
+    - ca_tanf_countable_vehicle_value # vehicle equity above the vehicle limit
