+++ policyengine_us/tests/policy/baseline/gov/states/ca/cdss/tanf/cash/eligibility/ca_tanf_eligible.yaml
@@ -48,7 +48,7 @@
   output:
     ca_tanf_eligible: false
 
-- name: Case 1, all eligible but vehicle value ineligible.
+- name: Case 1, all three conditions met.
   period: 2025
   absolute_error_margin: 0.1
   input:
@@ -56,18 +56,5 @@
     is_demographic_tanf_eligible: true
     ca_tanf_financial_eligible: true
     ca_tanf_resources_eligible: true
-    ca_tanf_vehicle_value_eligible: false
-  output:
-    ca_tanf_eligible: false
-
-- name: Case 2, all four conditions met.
-  period: 2025
-  absolute_error_margin: 0.1
-  input:
-    state_code: CA
-    is_demographic_tanf_eligible: true
-    ca_tanf_financial_eligible: true
-    ca_tanf_resources_eligible: true
-    ca_tanf_vehicle_value_eligible: true
   output:
     ca_tanf_eligible: true
