--- policyengine_us/tests/policy/baseline/gov/states/ca/cdss/tanf/cash/eligibility/ca_tanf_vehicle_value_eligible.yaml
@@ -1,91 +1,0 @@
-- name: 2023 Vehicle value eligible
-  period: 2023
-  input:
-    state_code: CA
-    household_vehicles_value: 32_045
-  output:
-    ca_tanf_vehicle_value_eligible: true
-
-- name: 2023 Vehicle value ineligible
-  period: 2023
-  input:
-    state_code: CA
-    household_vehicles_value: 32_046
-  output:
-    ca_tanf_vehicle_value_eligible: false
-
-- name: 2024 Vehicle value eligible
-  period: 2025
-  input:
-    state_code: CA
-    household_vehicles_value: 32_968
-  output:
-    ca_tanf_vehicle_value_eligible: true
-
-- name: 2024 Vehicle value ineligible
-  period: 2025
-  input:
-    state_code: CA
-    household_vehicles_value: 32_969
-  output:
-    ca_tanf_vehicle_value_eligible: false
-
-- name: Case 1, zero vehicle value always eligible.
-  period: 2025
-  absolute_error_margin: 0.1
-  input:
-    state_code: CA
-    household_vehicles_value: 0
-  output:
-    ca_tanf_vehicle_value_eligible: true
-
-- name: Case 2, historical 2023 vehicle limit at exact boundary $32,045.
-  period: 2024
-  absolute_error_margin: 0.1
-  input:
-    state_code: CA
-    household_vehicles_value: 32_045
-  output:
-    ca_tanf_vehicle_value_eligible: true
-
-- name: Case 3, historical 2023 vehicle limit one dollar over $32,045.
-  period: 2024
-  absolute_error_margin: 0.1
-  input:
-    state_code: CA
-    household_vehicles_value: 32_046
-  output:
-    ca_tanf_vehicle_value_eligible: false
-
-- name: Case 4, 2026 uses Jul 2025 vehicle limit at exact boundary $33,499.
-  period: 2026
-  input:
-    state_code: CA
-    household_vehicles_value: 33_499
-  output:
-    ca_tanf_vehicle_value_eligible: true
-
-- name: Case 5, 2026 uses Jul 2025 vehicle limit one dollar over $33,499.
-  period: 2026
-  input:
-    state_code: CA
-    household_vehicles_value: 33_500
-  output:
-    ca_tanf_vehicle_value_eligible: false
-
-- name: Case 6, 2027 uses Jul 2026 vehicle limit at exact boundary $33,626.
-  period: 2027
-  input:
-    state_code: CA
-    household_vehicles_value: 33_626
-  output:
-    ca_tanf_vehicle_value_eligible: true
-
-- name: Case 7, 2027 uses Jul 2026 vehicle limit one dollar over $33,626.
-  period: 2027
-  input:
-    state_code: CA
-    household_vehicles_value: 33_627
-  output:
-    ca_tanf_vehicle_value_eligible: false
-
